api/requirements.txt
websocket-client==0.57.0
inflection==0.5.0
Flask-SocketIO==4.3.1
Flask-Sockets==0.2.1
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
gunicorn==20.0.4
